Hardware: drop system subcommand and rename chassis to system
System subcommand displayed only hostname, which in case of hwinfo might be
pointless. After this patch, the hostname is displayed at the top of every
subcommand.

Chassis subcommand was renamed to system, because this name better represents
provided data - the data about "system", like the type (server, notebook), its
manufacturer and model.