service: redone interface
Column headers were quite confusing for both listing and show commands.
New they contain more human-friendly values and have proper names.

Removed oneshot option. This argument was out of false understanding
of EnabledDefault property. When it is set to NotApplicable, it means
the service is static - it can not be enabled or disabled. It is run
when needed (when something depends on it). Thus this options is removed
and Enabled column shows Static value for such services.

Stripped .service suffix from service names.

Resolves trac tickets: #239 #241