added smart hostname fallback finder
When no hosts were given to lmi on command line, it automatically chosen
'localhost' as a fallback. This is desirable behaviour for non-root
users because the selected protocol is https with SSL certificate
checks. And because pegasus's certificate is usually generated for
propert hostname and not for 'localhost', the connection will fail due
to certificate's unsuccessful validation.

This change brings simple algorithm that tries to find the best hostname
by trying to translate *fqdn*, *hostname* and *localhost* to IP in this
order. First name successfully translated will be chosen.